Enable global namespace and repository
The current documentation lists `namespace` and `repository` as global keys that would enable users to set the default namespace and repository for releases, but including these keys has no effect. This commit enables these keys to mirror the documentation and override the current defaults ('kube-system' for namespace and 'stable' for repository).