Python 3 Support (#84)
Hats off to @bambash for kicking this off!

This migration is intended to enable future deprecation of python 2.7 but also supporting python 3.6 & 3.7.

This PR:

Enables pytest being run on python 2.7, 3.6 and 3.7
Changes in python code to support both 2 and 3
This is just adding visibility to non-python3 changes going forward in the library. It will also fail if something is added that breaks python2, but the idea is that we should move away from that asap.