Supporting Schema Validation
Phase 1: Support validating the schema so that users
         get notified when there is an extra param or
         if they have an incorrect key/value pair in
         their course.yml.

I have housed the logic in a new module that reads the yaml and also
returns the correct errors expected at the cli.py. Earlier I was
planning on integrating this into course.py, as there's where the
initial YAML read happens, but since it's a part of __init__ in that
class, it makes loads of extra work for mocking. I opted to move this
ealier in the flow and have it fail faster even if we have to read the
course.yml twice. It's now integrated as the first thing in the cli.py.