Adding Diff command
This adds `diff` to the list of commands. It compares the output of `template`  and  `get_manifests`  using `difflib` . Whereas most other commands return a `HelmCommandResponse` response, that was not really work able without major re-write here because there is only a single instance of `self.response`  in the  class. Calling `template`  and  `get_manifests` in succession from an instnace of the Chart class overwrites the previous command's `self.response` . Less than ideal, but I created hidden methods to return the content of those commands that are then called by the new `diff`  command and I return the string representation of the diff as `self.resposne`  but it is not an instance of HelmResponse becaue it is not, technically, a HelmResponse.