Adding Diff command (#217)
* Adding Diff command

This adds `diff` to the list of commands. It compares the output of `template`  and  `get_manifests`  using `difflib` . Whereas most other commands return a `HelmCommandResponse` response, that was not really work able without major re-write here because there is only a single instance of `self.response`  in the  class. Calling `template`  and  `get_manifests` in succession from an instnace of the Chart class overwrites the previous command's `self.response` . Less than ideal, but I created hidden methods to return the content of those commands that are then called by the new `diff`  command and I return the string representation of the diff as `self.resposne`  but it is not an instance of HelmResponse becaue it is not, technically, a HelmResponse.

* addeding update path so that only releases that are different get run

* working around a logging thing I created

* adding docs

* adding e2e tests

* Update docs/README.md

Co-authored-by: Andrew Suderman <andrew@sudermanjr.com>

* Apply suggestions from code review

Co-authored-by: Andrew Suderman <andrew@sudermanjr.com>

* significant rework of the diff functionality to allow per document diffing that wtil works with the update method. Alos added --no-update-repos option because it makes rapid testing easier and it may be useful some day. unit tests incoming, but wanted to commit to get some other people to stress test it in the mean time

* changes to config class to allow tests to pass with new update repos attribute

* adding config and cli tests

* adding unit tests for the new diff classes

* more tests

* clean up and linting

* adding list type to tests, fixing where lists have no metadata, fixing where annotations is present but empty

* set it to use the source comment to find the list for comparision because lists do not have names to use

* flipped default on "update_repos" so that it will update repos on diff

* added the same try catch to the other commands

Co-authored-by: Andrew Suderman <andrew@sudermanjr.com>