Change working directoy to that of the course file
This Fixes https://github.com/FairwindsOps/reckoner/issues/248 on the main fork. The file paths in the course files were relative to the working directory of the command. This changes the directoy for the entire reckoner run.

This does not affect the hooks because those are run at a specific path. I chose to use `os.chdir` here instead of the `path` argument on the `call()` method because this was the simplest way with the least about of code change. To pass `path` through would require a change to many many lines of code.