Fix 264 - Env Var in context for diff
This PR fixes the issue where `diff` was not respecting the environment variable in the context.
It was due to the bespoke argument list in the `get_manifests` method. I added `_check_env_vars` there to replace the
envvars in the bespoke list. This was done by the `_pre_command` method already, but since the new arg list was created _after_,
it needs to be run again, or something like it.

Fixes #264