Fix Dependency update path (#327)
This PR fixes an issue with dependency update for a local chart when the `reckoner` command is not run in the same directory as the `course.yml` file. The update was being excluded because thecode checking if a chart existed was relative to the command directory and not the the `course.yml`. This changes the chart path check to be absolute _and_ uses the `course.yml` location as it's starting point.

Fixes #325