Raise error when command encounters one
This PR changes the error handling behavior of all commands.
Now, instead of reporting errors only at the CLI level, this fails fast by raising error when it is encountered.
For most commands the end result is the same but errors encountered for template, manifest, and consequently diff, were being ignored becaue of the way errors are propagated.
This fixes that

Fixes: #333