Adding init hook (#330)
* Adding init hook

Adding a hook that runs immediately after parsing the hooks from the course file. This means it will be run before _any_ of the commands instead of just the `plot` command. I think this is useful so that one can use a hook to initialize kubectl or authenticateinto a cluster for later commands.

* adding docs

Co-authored-by: Andrew Suderman <andrew@sudermanjr.com>